Add empty cart dialog

+++ app/src/main/java/com/tosh/poolandroid/view/fragment/CartFragment.kt
@@ -1,33 +1,42 @@
 package com.tosh.poolandroid.view.fragment
 
 
+import android.app.AlertDialog
 import android.content.Context
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.View.GONE
 import android.view.View.VISIBLE
 import android.view.ViewGroup
+import android.widget.Toast
 import androidx.activity.OnBackPressedCallback
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.bumptech.glide.request.transition.Transition
 
 import com.tosh.poolandroid.R
+import com.tosh.poolandroid.model.database.CartItemEntity
 import com.tosh.poolandroid.model.database.MainDatabase
 import com.tosh.poolandroid.view.activity.MainActivity
 import com.tosh.poolandroid.view.adapter.CartAdapter
 import com.tosh.poolandroid.viewmodel.MainViewModel
 import kotlinx.android.synthetic.main.fragment_cart.*
-import kotlinx.android.synthetic.main.fragment_placeholder.*
 import kotlinx.coroutines.launch
 
 class CartFragment : BaseFragment() {
 
     private var mainViewModel: MainViewModel? = null
+    lateinit var adapter: CartAdapter
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
         return inflater.inflate(R.layout.fragment_cart, container, false)
     }
 
@@ -36,11 +45,12 @@ class CartFragment : BaseFragment() {
 
         mainViewModel = ViewModelProviders.of(this).get(MainViewModel::class.java)
 
-        val adapter = CartAdapter()
+        adapter = CartAdapter()
         adapter.notifyDataSetChanged()
 
         fetchDataFromCart()
-        deleteCartItems()
+        deleteAllCartItems()
+        deleteSingleCartItem()
     }
 
     override fun onAttach(context: Context) {
@@ -55,6 +65,10 @@ class CartFragment : BaseFragment() {
         requireActivity().onBackPressedDispatcher.addCallback(this, callback)
     }
 
+    override fun onDetach() {
+        super.onDetach()
+        cartEmpty.visibility = GONE
+    }
 
     private fun fetchDataFromCart() {
 
@@ -65,8 +79,8 @@ class CartFragment : BaseFragment() {
             layoutManager = LinearLayoutManager(context)
         }
 
-        mainViewModel!!.getCartTotal().observe(viewLifecycleOwner, Observer {total ->
-           val grandTotal = total.toInt()
+        mainViewModel!!.getCartTotal().observe(viewLifecycleOwner, Observer { total ->
+            val grandTotal = total.toInt()
             totalCart.text = "Total $grandTotal KES"
         })
 
@@ -87,18 +101,36 @@ class CartFragment : BaseFragment() {
 
     }
 
-    private fun deleteCartItems() {
+    fun deleteSingleCartItem() {
+
+    }
+
+
+    private fun deleteAllCartItems() {
         val placeholderFragment = CartFragment()
+        val emptyDialog = AlertDialog.Builder(context)
+        emptyDialog.setTitle("Empty Cart")
+        emptyDialog.setMessage("Are you sure ?")
         emptyCart.setOnClickListener {
-            launch {
-                context.let {
-                    MainDatabase.getInstance(it!!)!!.cartItemDao().deleteCart()
-                    val transaction = activity!!.supportFragmentManager.beginTransaction()
-                    transaction.replace(R.id.details_fragment, placeholderFragment)
-                    transaction.commit()
+            emptyDialog.setPositiveButton(android.R.string.yes) { _, _ ->
+                launch {
+                    context.let {
+                        MainDatabase.getInstance(it!!)!!.cartItemDao().deleteCart()
+                        val transaction = activity!!.supportFragmentManager.beginTransaction()
+                        transaction.replace(R.id.details_fragment, placeholderFragment)
+                        transaction.commit()
+                    }
                 }
             }
+
+            emptyDialog.setNegativeButton(android.R.string.no) { _, _ ->
+            }
+            emptyDialog.show()
         }
     }
+
+//    override fun onItemClick(position: Int) {
+//        Log.e("Yezzz", "Clicked")
+//    }
 }
